IOTAPP/requirements.txt
Flask==3.1.0
flask_sqlalchemy==3.1.1
python-dotenv==1.1.0
Requests==2.32.3
SQLAlchemy==2.0.36
Werkzeug==3.1.3
asposestorage==1.0.2
config==0.5.1
Flask==3.1.0
Requests==2.32.3
SQLAlchemy==2.0.36
utils==1.0.2
Werkzeug==3.1.3

#requirements neded 